Starting to split contracts into recurring_contract/compassion_contract
