+++ contract_compassion/__init__.py
@@ -3,19 +3,10 @@
 #
 #    Copyright (C) 2014 Compassion CH (http://www.compassion.ch)
 #    Releasing children from poverty in Jesus' name
-#    @author: Cyril Sester, Emanuel Cino
+#    @author: David Coninckx <david@coninckx.com>
 #
 #    The licence is in the file __openerp__.py
 #
 ##############################################################################
 
-# from . import contract_group
-from . import contracts
-from . import invoice_line
-from . import res_partner
-from . import product
-from . import move_line
-# from . import contract_origin
-from . import account_banking_mandate
-from . import project_compassion
-from . import child_compassion
+from . import model
