+++ contract_compassion/view/contract_origin_view.xml
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+<data>
+	<record id="view_contract_origin_tree" model="ir.ui.view">
+		<field name="name">contract.origin.tree</field>
+		<field name="model">recurring.contract.origin</field>
+		<field name="arch" type="xml">
+			<tree string="Origins">
+                <field name="name"/>
+                <field name="type"/>
+                <field name="partner_id"/>
+                <field name="analytic_id"/>
+                <field name="country_id"/>
+                <field name="won_sponsorships"/>
+            </tree>
+		</field>
+	</record>
+	
+	<record id="view_contract_origin_form" model="ir.ui.view">
+		<field name="name">contract.origin.form</field>
+		<field name="model">recurring.contract.origin</field>
+		<field name="arch" type="xml">
+            <form string="Origin" version="7.0">
+                <sheet>
+                    <div class="oe_title">
+                        <group>
+                            <field name="name"/>
+                            <field name="type"/>
+                        </group>
+                    </div>
+                    <separator />
+                    <group>
+                        <field name="partner_id" attrs="{'required': [('type', '=', 'partner')]}"
+                               options="{'create': false, 'm2o_dialog': false}"/>
+                        <field name="analytic_id" attrs="{'invisible': [('type', 'not in', ('marketing','partner'))],
+                                                          'required': [('type', '=', 'marketing')]}"
+                               domain="[('parent_id.name','=','Campaign')]"/>
+                        <field name="country_id" attrs="{'invisible': [('type', '!=', 'transfer')],
+                                                         'required': [('type','=','transfer')]}"/>
+                        <field name="other_name" attrs="{'invisible': [('type', '!=', 'other')]}"/>
+                        <field name="won_sponsorships"/>
+                    </group>
+                    <notebook>
+                        <page string="Originated Sponsorships">
+                            <field name="contract_ids" attrs="{'invisible': [('contract_ids', '=', False)]}"/>
+                        </page>
+                    </notebook>
+                </sheet>
+                
+            </form>
+		</field>
+	</record>
+    
+    <record id="view_recurring_contract_filter" model="ir.ui.view">
+		<field name="name">recurring.contract.origin.select</field>
+		<field name="model">recurring.contract.origin</field>
+		<field name="arch" type="xml">
+            <search string="Search origin">
+                <field name="name"/>
+                <field name="type"/>
+                <group expand="0" string="Group By...">
+                    <filter string="Type" domain="[]" context="{'group_by':'type'}"/>
+                </group>
+            </search>
+        </field>
+    </record>
+    
+    <record id="open_view_origin_tree" model="ir.actions.act_window">
+        <field name="name">Origins</field>
+        <field name="res_model">recurring.contract.origin</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form,tree</field>
+        <field name="view_id" ref="view_contract_origin_tree"/>
+    </record>
+    
+    <menuitem id="menu_sponsorship_origin" parent="child_compassion.menu_sponsorship_section"
+              name="Origins" groups="child_compassion.group_sponsorship" sequence="4"
+              action="open_view_origin_tree"/>
+</data>
+</openerp>
