+++ crm_compassion/view/contract_origin_view.xml
@@ -49,7 +49,7 @@
         <record id="view_recurring_contract_filter" model="ir.ui.view">
             <field name="name">recurring.contract.select</field>
             <field name="model">recurring.contract</field>
-            <field name="inherit_id" ref="sponsorship_compassion.view_recurring_contract_compassion_filter" />
+            <field name="inherit_id" ref="contract_compassion.view_recurring_contract_compassion_filter" />
             <field name="arch" type="xml">
                 <data>
                     <field name="origin_id" position="after">
@@ -68,13 +68,13 @@
             </field>
         </record>
 
-        <record id="view_contract_tree_crm" model="ir.ui.view">
-            <field name="name">recurring.contract.crm.tree</field>
-            <field name="model">recurring.contract</field>
-            <field name="inherit_id" ref="sponsorship_compassion.view_recurring_contract_tree_compassion"/>
-            <field eval="18" name="priority"/>
-            <field name="arch" type="xml">
-                <data>
+    <record id="view_contract_tree_crm" model="ir.ui.view">
+			<field name="name">recurring.contract.crm.tree</field>
+			<field name="model">recurring.contract</field>
+			<field name="inherit_id" ref="contract_compassion.view_recurring_contract_tree_compassion"/>
+			<field eval="18" name="priority"/>
+			<field name="arch" type="xml">
+				<data>
                     <field name="origin_id" position="after">
                         <field name="user_id" />
                     </field>
