+++ message_center_compassion/view/contracts_view.xml
@@ -10,7 +10,7 @@
         <record id="view_gmc_contract_filter" model="ir.ui.view">
             <field name="name">gmc.contract.filter</field>
             <field name="model">recurring.contract</field>
-            <field name="inherit_id" ref="sponsorship_compassion.view_recurring_contract_compassion_filter" />
+            <field name="inherit_id" ref="contract_compassion.view_recurring_contract_compassion_filter" />
             <field name="arch" type="xml">
                 <data>
                     <filter name="finished" position="after">
